Forgot to include the sysdata.Rda file
Without it automated checking and deployment fail.

I left it out originally as I don't like keeping binaries in git if I don't have to but I've realised that leaving it out breaks my travis build and would break any devtools::install_github's as well, so I've put it in.