Added an additional script to clean the lists
It's easy for long files like adjectives.txt and nouns.txt to become full of duplicates and unwanted characters and so on over time.

clean_list.sh checks for non-lower-case alpha characters, sorts and removes duplicate entries in order to keep the list nice and tidy.

This makes it easier for a human to maintain and also helps to ensure that the package tests pass.

Creating this file also led to automatic updates to nouns.txt, adjectives.txt and sysdata.R. The Makefile was updated to include the new script.